gov2 refactor the summary components
