+++ packages/next-common/components/summary/gov2Summary.js
@@ -2,50 +2,17 @@
 
 import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import styled from "styled-components";
 import { currentNodeSelector } from "../../store/reducers/nodeSlice";
 import useApi from "../../utils/hooks/useApi";
-import { SecondaryCard } from "../styled/containers/secondaryCard";
 import Content from "./cardContent";
 import nextApi from "../../services/nextApi";
 import { gov2ReferendumsSummaryApi } from "../../services/url";
-
-const Wrapper = styled.div`
-  display: flex;
-
-  > :not(:first-child) {
-    margin-left: 16px;
-  }
-
-  @media screen and (max-width: 768px) {
-    flex-direction: column;
-    > :not(:first-child) {
-      margin-left: 0;
-      margin-top: 16px;
-    }
-  }
-`;
-
-const Card = styled(SecondaryCard)`
-  position: relative;
-  color: ${(props) => props.theme.textPrimary};
-  flex: 0 1 33.33%;
-  height: 88px;
-`;
-
-const Title = styled.div`
-  font-weight: bold;
-  font-size: 12px;
-  line-height: 100%;
-  letter-spacing: 0.16em;
-  color: ${(props) => props.theme.textTertiary};
-  margin-bottom: 8px;
-  text-transform: uppercase;
-`;
-
-const GreyText = styled.span`
-  color: ${(props) => props.theme.textTertiary}; !important;
-`;
+import {
+  SummaryCard,
+  SummaryGreyText,
+  SummaryTitle,
+  SummaryWrapper,
+} from "./styled";
 
 export default function Gov2Summary({ chain }) {
   const [summary, setSummary] = useState({});
@@ -63,28 +30,28 @@ export default function Gov2Summary({ chain }) {
   }, [chain, api]);
 
   return (
-    <Wrapper>
-      <Card>
-        <Title>Confirming</Title>
+    <SummaryWrapper>
+      <SummaryCard>
+        <SummaryTitle>Confirming</SummaryTitle>
         <Content>
           <span>{summary.confirmingCount || 0}</span>
         </Content>
-      </Card>
-      <Card>
-        <Title>Deciding</Title>
+      </SummaryCard>
+      <SummaryCard>
+        <SummaryTitle>Deciding</SummaryTitle>
         <Content>
           <span>{summary.decidingCount || 0}</span>
         </Content>
-      </Card>
-      <Card>
-        <Title>Active</Title>
+      </SummaryCard>
+      <SummaryCard>
+        <SummaryTitle>Active</SummaryTitle>
         <Content>
           <span>
             {summary.activeCount || 0}
-            <GreyText> / {summary.total || 0}</GreyText>
+            <SummaryGreyText> / {summary.total || 0}</SummaryGreyText>
           </span>
         </Content>
-      </Card>
-    </Wrapper>
+      </SummaryCard>
+    </SummaryWrapper>
   );
 }
